requirements.txt
# Dependencias principales
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.1
jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1

# Autenticación y seguridad
pyjwt==2.1.0
python-dotenv==1.0.0

# Serialización y validación
marshmallow==3.13.0

# Servidor de producción
gunicorn==20.1.0

# Solicitudes HTTP
requests==2.31.0
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
urllib3==2.0.3

# Utilidades
python-dateutil==2.8.2
six==1.16.0
pytz==2023.3

# Pruebas
pytest==7.3.1
pytest-flask==1.2.0
